+++ release-notes.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Screencasts now support window capture.
+- Screencasts now support workspace capture.
 - Add support for wp-alpha-modifier.
 - Add support for per-device keymaps.
 - Add support for virtual-keyboard-unstable-v1.